+++ src/lib/components/user-card.svelte
@@ -1,11 +1,18 @@
 <script lang="ts">
 	import type { Prisma, Question } from '@prisma/client';
+	import { onMount } from 'svelte';
 	import SvelteMarkdown from 'svelte-markdown';
 
 	export let user: Partial<Prisma.UserGetPayload<{ include: { authUser: true; decision: true } }>>;
 	export let questions: Question[];
 
-	$: application = user.application as Record<string, unknown>;
+	// eslint-disable-next-line @typescript-eslint/no-explicit-any
+	const application = user.application as Record<string, any>;
+
+	let origin = '';
+	onMount(() => {
+		origin = location.origin;
+	});
 </script>
 
 <p><b>Role</b> {user.authUser?.roles.join(', ')}</p>
@@ -16,7 +23,22 @@
 </p>
 {#each questions as question}
 	<SvelteMarkdown source={question.label} />
-	<blockquote>{application[question.id]}</blockquote>
+	<blockquote>
+		{#if application[question.id] === undefined || application[question.id] === ''}
+			<i>No answer given</i>
+		{:else if question.type === 'FILE'}
+			<a
+				href={`/files/${user.authUserId}/${question.id}`}
+				target="_blank"
+				rel="noopener noreferrer"
+			>
+				{`${origin}/files/${user.authUserId}/${question.id}`}</a
+			>
+		{:else if Array.isArray(application[question.id])}
+			{application[question.id].join(', ')}
+		{:else}
+			{application[question.id]}{/if}
+	</blockquote>
 {/each}
 
 <style>