+++ src/routes/admin/+page.server.ts
@@ -28,29 +28,10 @@ export const actions = {
 		} catch (e) {
 			confirmBy = null;
 		}
-		const homepageText = formData.get('homepageText') as string;
-		const submitTemplate = formData.get('submitTemplate') as string;
-		const acceptTemplate = formData.get('acceptTemplate') as string;
-		const rejectTemplate = formData.get('rejectTemplate') as string;
-		const waitlistTemplate = formData.get('waitlistTemplate') as string;
-		const confirmTemplate = formData.get('confirmTemplate') as string;
-		const declineTemplate = formData.get('declineTemplate') as string;
 		await trpc(locals.auth).settings.update({
 			applicationOpen,
 			confirmBy,
-			homepageText,
-			submitTemplate,
-			acceptTemplate,
-			rejectTemplate,
-			waitlistTemplate,
-			confirmTemplate,
-			declineTemplate,
 		});
 		return 'Saved settings!';
-	},
-
-	release: async ({ locals }) => {
-		await trpc(locals.auth).admissions.releaseAllDecisions();
-		return 'Released all decisions!';
 	},
 };