+++ src/routes/admin/questions/+page.svelte
@@ -5,8 +5,6 @@
 	export let data;
 </script>
 
-<h1>Registration Questions</h1>
-
 <form
 	action="?/update"
 	method="POST"
@@ -50,18 +48,18 @@
 				</div>
 			</div>
 			<div>
-				<label for={question.id}><b>{question.type}</b> Label</label>
+				<label for={question.id + '_label'}><b>{question.type}</b> Label</label>
 				<input
 					value={question.label}
 					name={question.id + '_label'}
-					id={question.id}
+					id={question.id + '_label'}
 					placeholder="What is your name?"
 				/>
 			</div>
 			<!-- Question-type-specific fields -->
 			{#if question.type === 'SENTENCE' || question.type === 'PARAGRAPH'}
 				<div>
-					<label for={question.id}>Placeholder</label>
+					<label for={question.id + '_placeholder'}>Placeholder</label>
 					<input
 						value={question.placeholder}
 						name={question.id + '_placeholder'}
@@ -70,7 +68,7 @@
 					/>
 				</div>
 				<div>
-					<label for={question.id}>Response must match regex:</label>
+					<label for={question.id + '_regex'}>Response must match regex:</label>
 					<input
 						value={question.regex}
 						name={question.id + '_regex'}
@@ -80,7 +78,7 @@
 				</div>
 			{:else if question.type === 'NUMBER'}
 				<div>
-					<label for={question.id}>Placeholder</label>
+					<label for={question.id + '_placeholder'}>Placeholder</label>
 					<input
 						value={question.placeholder}
 						type="number"
@@ -91,7 +89,7 @@
 				</div>
 				<div class="flex-row">
 					<div>
-						<label for={question.id}>Minimum</label>
+						<label for={question.id + '_min'}>Minimum</label>
 						<input
 							value={question.min}
 							type="number"
@@ -102,7 +100,7 @@
 						/>
 					</div>
 					<div>
-						<label for={question.id}>Maximum</label>
+						<label for={question.id + '_max'}>Maximum</label>
 						<input
 							value={question.max}
 							type="number"
@@ -113,7 +111,7 @@
 						/>
 					</div>
 					<div>
-						<label for={question.id}>Step</label>
+						<label for={question.id + '_step'}>Step</label>
 						<input
 							value={question.step}
 							type="number"
@@ -124,15 +122,50 @@
 						/>
 					</div>
 				</div>
-			{:else if question.type === 'DROPDOWN' || question.type === 'MULTISELECT'}
+			{:else if question.type === 'DROPDOWN' || question.type === 'RADIO'}
 				<div>
-					<label for={question.id}>Options</label>
+					<label for={question.id + '_options'}>Options</label>
 					<textarea
 						value={question.options.join('\n')}
 						name={question.id + '_options'}
 						id={question.id + '_options'}
 						placeholder="Write one option per line, like this:&#13;OPTION 1&#13;OPTION 2&#13;OPTION 3"
 					/>
+				</div>
+				<div class="flex-row">
+					<Toggle
+						name={question.id + '_multiple'}
+						label="Allow multiple selections"
+						checked={Boolean(question.multiple)}
+					/>
+					<Toggle
+						name={question.id + '_custom'}
+						label="Allow custom response entry"
+						checked={Boolean(question.custom)}
+					/>
+				</div>
+			{:else if question.type === 'FILE'}
+				<div class="flex-row">
+					<div>
+						<label for={question.id + '_accept'}>Accepted types:</label>
+						<input
+							value={question.accept}
+							name={question.id + '_accept'}
+							id={question.id + '_accept'}
+							placeholder=".doc, .docx, .pdf"
+						/>
+					</div>
+					<div>
+						<label for={question.id + '_maxSizeMB'}>Max file size (MB)</label>
+						<input
+							value={question.maxSizeMB}
+							type="number"
+							name={question.id + '_maxSizeMB'}
+							id={question.id + '_maxSizeMB'}
+							placeholder="10"
+							step="any"
+						/>
+					</div>
 				</div>
 			{/if}
 		</fieldset>
@@ -144,7 +177,6 @@
 			<option value="PARAGRAPH">Paragraph</option>
 			<option value="NUMBER">Number</option>
 			<option value="DROPDOWN">Dropdown</option>
-			<option value="MULTISELECT">Multiselect</option>
 			<option value="CHECKBOX">Checkbox</option>
 			<option value="RADIO">Radio</option>
 			<option value="FILE">File</option>
@@ -194,7 +226,7 @@
 
 	.flex-row {
 		display: flex;
-		flex-direction: row;
+		flex-flow: row wrap;
 		justify-content: space-between;
 		gap: 1rem;
 	}